// src/plat/schemas/plat.entity.ts
//  Generated by entity-cli

import * as mongoose from "mongoose";


export const schemaName = 'Plat';

export interface PlatInterface extends mongoose.Document {
    nom: string;
    description: string;
    prix: number;
    image: string;
}

export const Plat = new mongoose.Schema(
    {
        nom: {
            type: String,
            require: true,
        },
        description: {
            type: String,
            require: true,
        },
        prix: {
            type: Number,
            require: true,
        },
        image: {
            type: String,
            require: true,
            default: "null"
        },
    },
    { timestamps: true },
);